Refactoring directory, nothing change
