+++ Assemblies/static_quality_plus.dll
