+++ Source/static_quality_plus/SpecialInjector.cs
@@ -1,0 +1,35 @@
+﻿
+using System.Linq;
+using System.Reflection;
+using Verse;
+
+namespace static_quality
+{
+    public class SpecialInjector
+    {
+
+        public virtual bool Inject()
+        {
+            Log.Error("This should never be called.");
+            return false;
+        }
+    }
+
+    [StaticConstructorOnStartup]
+    internal static class DetourInjectorSQ
+    {
+        private static Assembly Assembly { get { return Assembly.GetAssembly(typeof(DetourInjector)); } }
+        private static string AssemblyName { get { return Assembly.FullName.Split(',').First(); } }
+        static DetourInjectorSQ()
+        {
+            LongEventHandler.QueueLongEvent(Inject, "Initializing", true, null);
+        }
+
+        private static void Inject()
+        {
+            var injector = new DetourInjector();
+            if (injector.Inject()) Log.Message(AssemblyName + " injected.");
+            else Log.Error(AssemblyName + " failed to get injected properly.");
+        }
+    }
+}
