--- static_quality_plus/static_quality_plus.sln
@@ -1,22 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "static_quality_plus", "static_quality_plus.csproj", "{7409D1D7-73E3-4B85-AE95-54FFAA969FFD}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7409D1D7-73E3-4B85-AE95-54FFAA969FFD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7409D1D7-73E3-4B85-AE95-54FFAA969FFD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7409D1D7-73E3-4B85-AE95-54FFAA969FFD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7409D1D7-73E3-4B85-AE95-54FFAA969FFD}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
